blk-mq: update hardware and software queues for sleeping alloc
If we end up sleeping due to running out of requests, we should update
the hardware and software queues in the map ctx structure. Otherwise we
could end up having rq->mq_ctx point to the pre-sleep context, and risk
corrupting ctx->rq_list since we'll be grabbing the wrong lock when
inserting the request.

Reported-by: Dave Jones <davej@codemonkey.org.uk>
Reported-by: Chris Mason <clm@fb.com>
Tested-by: Chris Mason <clm@fb.com>
Fixes: 63581af3f31e ("blk-mq: remove non-blocking pass in blk_mq_map_request")
Signed-off-by: Jens Axboe <axboe@fb.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>
Signed-off-by: djb77 <dwayne.bakewell@gmail.com>